+++ scripts/list.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-# List all crawled sites
-# Usage: ./list.sh
-
-python3 -m doc_search list